Separate logic from data model.
